% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_triangular_funcs.R
\name{.diag_locs}
\alias{.diag_locs}
\title{Indices of diagonal elements when in lower triangular vector}
\usage{
.diag_locs(r)
}
\arguments{
\item{r}{An integer. Dimension of matrix}
}
\description{
Indices of diagonal elements when in lower triangular vector
}
\keyword{internal}
