% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.standardize}
\alias{.standardize}
\title{Standardize data}
\usage{
.standardize(x)
}
\arguments{
\item{x}{data vector}
}
\value{
data with mean 0 and std 1
}
\description{
Standardize data
}
\keyword{internal}
