% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformZi.R
\name{.transformZi}
\alias{.transformZi}
\title{Transform Zi for RVB2}
\usage{
.transformZi(Zi)
}
\arguments{
\item{Zi}{A vector or matrix}
}
\value{
(Zi^T Zi)^{-1} Zi
}
\description{
Transform Zi for RVB2
}
\keyword{internal}
