% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_rvb1.R
\name{.ZgH}
\alias{.ZgH}
\title{Precompute for RVB1.}
\usage{
.ZgH(y, Z, model, etahat, m, n, r)
}
\arguments{
\item{y}{List of numeric vectors. y[\link{i}] is the response vector for cluster i.}

\item{Z}{List of vectors (r = 1) or matrices. Z[\link{i}] is the Z matrix for cluster i.}

\item{model}{Character. Binomial or Poisson model}

\item{etahat}{List of numeric vectors. etahat[\link{i}] is the MLE for cluster i}

\item{m}{Numeric. Number of trials in Binomial model.}

\item{n}{Numeric. Number of clusters.}

\item{r}{Numeric. Number of local variables (dimensionality of random effects).}
}
\value{
List of precomputed vals
}
\description{
See lambda_i in equation 7 of Tan (2021).
}
\keyword{internal}
