% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherscoring.R
\name{fisherscoring}
\alias{fisherscoring}
\title{Fisher scoring for RVB2}
\usage{
fisherscoring(etahati, Zyi, Zi, Zti, Xibeta, Omega, model, mi)
}
\arguments{
\item{etahati}{Vector. MLE under the regularized model from RVB1}

\item{Zyi}{Numeric or vector. Zi^T yi}

\item{Zi}{Vector. Covariates for local variables}

\item{Zti}{Vector. Output from transformZi}

\item{Xibeta}{Vector. Xi^T beta}

\item{Omega}{Matrix. Local variable prior precision matrix}

\item{model}{Character. Either "poisson" or "binomial"}

\item{mi}{Numeric. Number of trials if binomial}
}
\value{
A numeric or vector representing the mode of the conditional posterior
density of the local variables bi given the global variables and data.
}
\description{
Fisher scoring for RVB2
}
