% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_triangular_funcs.R
\name{get_lower_tri_vector}
\alias{get_lower_tri_vector}
\title{Vector from lower triangular matrix entries}
\usage{
get_lower_tri_vector(W)
}
\arguments{
\item{W}{an n x n matrix}
}
\value{
Vector of size n(n+1)/2 containing lower triangular matrix entries.
}
\description{
Extract the lower triangular matrix entries into a vector, stacking columns left to right.
}
