% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLB.R
\name{GLB}
\alias{GLB}
\title{Gradient of least squares regression line}
\usage{
GLB(j, tol, par)
}
\arguments{
\item{j}{Iteration in thousands.}

\item{tol}{Iteration tolerance}

\item{par}{parameter}
}
\value{
Gradient of least squares regression line fitted to the last means.
}
\description{
Gradient of least squares regression line
}
