% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_var_components.R
\name{global_var_components}
\alias{global_var_components}
\title{Get global variable density components}
\usage{
global_var_components(theta, n, p, r)
}
\arguments{
\item{theta}{A vector of length p + r(r+1)/2 + nr, containing first local then global variables}

\item{n}{Integer. Number of total observations in data.}

\item{p}{Integer. Dimension of fixed effects (beta) in a GLMM}

\item{r}{Integer. Dimension of random effects (b) in a GLMM}
}
\value{
List containing components needed to compute log joint density. In reverse order,
includes the precision matrix Omega, its Cholesky decomposition W, the vector of fixed
effects beta, the weighted sum of the log diagonal of W, and the sequence used to determine
the weights.
}
\description{
Obtains log joint density components for global variables, to be used when
computing the log joint density. Values are due to the default conjugate
Wishart prior from Kass and Natarajan (2006). See Tan (2021) for more details.
}
