% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_partitions_and_derivs.R
\name{h0}
\alias{h0}
\title{Log partition function}
\usage{
h0(model, x, m)
}
\arguments{
\item{model}{Character. Either "binomial" or "poisson"}

\item{x}{Numeric. Probability for binomial, mean for poisson}

\item{m}{Integer. Count for binomial}
}
\value{
Log partition function for the exponential family
}
\description{
Log partition function
}
