% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_partitions_and_derivs.R
\name{h1}
\alias{h1}
\title{First derivative of log partition}
\usage{
h1(model, x, m)
}
\arguments{
\item{model}{Character. Either "binomial" or "poisson"}

\item{x}{Numeric. Probability for binomial, mean for poisson}

\item{m}{Integer. Count for binomial}
}
\value{
First derivative of log partition function.
}
\description{
First derivative of log partition
}
