% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_partitions_and_derivs.R
\name{h2}
\alias{h2}
\title{Second derivative of log partition}
\usage{
h2(model, x, m)
}
\arguments{
\item{model}{Character. Either "binomial" or "poisson"}

\item{x}{Numeric. Probability for binomial, mean for poisson}

\item{m}{Integer. Count for binomial}
}
\value{
Second derivative of log partition function.
}
\description{
Second derivative of log partition
}
