% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_partitions_and_derivs.R
\name{h3}
\alias{h3}
\title{Third derivative of log partition}
\usage{
h3(model, x, m)
}
\arguments{
\item{model}{Character. Either "binomial" or "poisson"}

\item{x}{Numeric. Probability for binomial, mean for poisson}

\item{m}{Integer. Count for binomial}
}
\value{
Third derivative of log partition function.
}
\description{
Third derivative of log partition
}
