% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_for_posterior_covariance.R
\name{indices_for_posterior_covariance}
\alias{indices_for_posterior_covariance}
\title{Indices for posterior covariance}
\usage{
indices_for_posterior_covariance(n, r, p)
}
\arguments{
\item{n}{Integer. Number of subjects/clusters.}

\item{r}{Integer. Dimensionality of random effects}

\item{p}{Integer. Dimensionality of fixed effects}
}
\value{
List containing row and column indices for non-zero elements.
}
\description{
Finds the indices for the non-zero elements in the Cholesky decomposition
of the posterior covariance matrix. The posterior covariance is assumed to
to be a block matrix of corresponding to the local and global variables
}
