% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNprior.R
\name{KNprior}
\alias{KNprior}
\title{Kass and Natarajan prior for random effects precision matrix}
\usage{
KNprior(model, pred, Z, m = 1)
}
\arguments{
\item{model}{Character. Either "poisson" or "binomial"}

\item{pred}{Vector. GLM predictions in y-scale}

\item{Z}{List. Covariates for random effects}

\item{Integer.}{Number of trials if model = "binomial". If "poisson", leave as 1.}
}
\value{
List containing Wishart prior parameter nu, S-inverse, and S
}
\description{
Kass and Natarajan prior for random effects precision matrix
}
