% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB_RVB1.R
\name{LB_RVB1}
\alias{LB_RVB1}
\title{Evidence Lower Bound for RVB1}
\usage{
LB_RVB1(mu, C, y, X, Z, Zg, ZH, vbeta0, Sinv, model, m, n, p, r, d, N = 1000)
}
\arguments{
\item{mu}{Vector. Mean in variational approximation}

\item{C}{Lower triangular matrix. Cholesky decomposition of variance}

\item{y}{List. Responses per cluster}

\item{X}{List. Covariates per cluster for fixed effects}

\item{Z}{List. Covariates per cluster for random effects}

\item{Zg}{Vector or Matrix. Precomputed value}

\item{ZH}{List. Precomputed value}

\item{vbeta0}{Numeric. Prior variance for fixed effects}

\item{Sinv}{Matrix. Inverse of prior covariance matrix for random effects}

\item{model}{Character. Either "poisson" or "binomial"}

\item{m}{Integer. Number of trials if "binomial"}

\item{n}{Integer. Number of clusters}

\item{p}{Integer. Number of fixed effects}

\item{r}{Integer. Number of random effects}

\item{N}{Integer. Number of iterations for ELBO average.}

\item{ttheta}{Numeric vector. All global and local variables.}
}
\value{
Value of the evidence lower bound based on RVB1 optimization
}
\description{
Evidence Lower Bound for RVB1
}
