% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LB_RVB2.R
\name{LB_RVB2}
\alias{LB_RVB2}
\title{Evidence lower bound for RVB2}
\usage{
LB_RVB2(
  mu,
  C,
  y,
  X,
  Z,
  Zt,
  Zy,
  etahat,
  vbeta0,
  Sinv,
  model,
  m,
  n,
  p,
  r,
  d,
  N = 1000
)
}
\arguments{
\item{mu}{Vector. Mean in variational approximation}

\item{C}{Lower triangular matrix. Cholesky decomposition of variance}

\item{y}{List. Responses per cluster}

\item{X}{List. Covariates per cluster for fixed effects}

\item{Z}{List. Covariates per cluster for random effects}

\item{Zt}{List. Output from transformZi}

\item{Zy}{Numeric or vector. Z^T y}

\item{etahat}{List. Contains per cluster MLEs under RVB1 regularized model}

\item{vbeta0}{Numeric. Prior variance for fixed effects}

\item{Sinv}{Matrix. Inverse of prior covariance matrix for random effects}

\item{model}{Character. Either "poisson" or "binomial"}

\item{m}{Integer. Number of trials if "binomial"}

\item{n}{Integer. Number of clusters}

\item{p}{Integer. Number of fixed effects}

\item{r}{Integer. Number of random effects}

\item{d}{Integer. Total number of variables}

\item{N}{Integer. Number of iterations for ELBO average}
}
\value{
Value of the evidence lower bound based on RVB2 optimization
}
\description{
Evidence lower bound for RVB2
}
