% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lgrad_RVB1.R
\name{Lgrad_RVB1}
\alias{Lgrad_RVB1}
\title{Log joint density and gradient for RVB1}
\usage{
Lgrad_RVB1(ttheta, y, X, Z, Zg, ZH, vbeta0, Sinv, model, n, p, r, d)
}
\arguments{
\item{ttheta}{Numeric vector. All global and local variables.}

\item{y}{List. Responses per cluster}

\item{X}{List. Covariates per cluster for fixed effects}

\item{Z}{List. Covariates per cluster for random effects}

\item{Zg}{Vector or Matrix. Precomputed value}

\item{ZH}{List. Precomputed value}

\item{vbeta0}{Numeric. Prior variance for fixed effects}

\item{Sinv}{Matrix. Inverse of prior covariance matrix for random effects}

\item{model}{Character. Either "poisson" or "binomial"}

\item{n}{Integer. Number of clusters}

\item{p}{Integer. Number of fixed effects}

\item{r}{Integer. Number of random effects}

\item{d}{Integer. Total number of local and global variables.}
}
\value{
List including Log joint density and gradient of log joint density under RVB1
}
\description{
Log joint density and gradient for RVB1
}
