% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lgrad_RVB2.R
\name{Lgrad_RVB2}
\alias{Lgrad_RVB2}
\title{Log joint density and gradient for RVB2}
\usage{
Lgrad_RVB2(ttheta, y, X, Z, Zt, Zy, etahat, vbeta0, Sinv, model, m, n, p, r, d)
}
\arguments{
\item{ttheta}{Numeric vector. All global and local variables.}

\item{y}{List. Responses per cluster}

\item{X}{List. Covariates per cluster for fixed effects}

\item{Z}{List. Covariates per cluster for random effects}

\item{Zt}{List. Output from transformZi}

\item{Zy}{Numeric or vector. Z^T y}

\item{etahat}{List. Contains per cluster MLEs under RVB1 regularized model}

\item{vbeta0}{Numeric. Prior variance for fixed effects}

\item{Sinv}{Matrix. Inverse of prior covariance matrix for random effects}

\item{model}{Character. Either "poisson" or "binomial"}

\item{m}{Integer. Number of trials if "binomial"}

\item{n}{Integer. Number of clusters}

\item{p}{Integer. Number of fixed effects}

\item{r}{Integer. Number of random effects}

\item{d}{Integer. Total number of variables}
}
\value{
List. Containing log joint density and the gradient vector
for the log joint density under RVB2
}
\description{
Log joint density and gradient for RVB2
}
