% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpy_RVB1.R
\name{logpy_RVB1}
\alias{logpy_RVB1}
\title{Log joint density under RVB1}
\usage{
logpy_RVB1(ttheta, y, X, Z, Zg, ZH, vbeta0, Sinv, model, m, n, p, r)
}
\arguments{
\item{ttheta}{Numeric vector. All global and local variables.}

\item{y}{List. Responses per cluster}

\item{X}{List. Covariates per cluster for fixed effects}

\item{Z}{List. Covariates per cluster for random effects}

\item{Zg}{Vector or Matrix. Precomputed value}

\item{ZH}{List. Precomputed value}

\item{vbeta0}{Numeric. Prior variance for fixed effects}

\item{Sinv}{Matrix. Inverse of prior covariance matrix for random effects}

\item{model}{Character. Either "poisson" or "binomial"}

\item{m}{Integer. Number of trials if "binomial"}

\item{n}{Integer. Number of clusters}

\item{p}{Integer. Number of fixed effects}

\item{r}{Integer. Number of random effects}
}
\value{
Log joint density under RVB1 model reparameterization
}
\description{
Log joint density under RVB1
}
