% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postb_RVB1.R
\name{postb_RVB1_b1}
\alias{postb_RVB1_b1}
\title{Sample from posterior under RVB1}
\usage{
postb_RVB1_b1(RVB, y, X, Z, etahat, N, L, model, m = 1)
}
\arguments{
\item{RVB}{List. Output from Alg_RVB1}

\item{y}{List. Responses per cluster}

\item{X}{List. Covariates per cluster for fixed effects}

\item{Z}{List. Covariates per cluster for random effects#'}

\item{etahat}{List. Estimate of canonical parameter about which to approximate}

\item{N}{Integer. Sample size}

\item{L}{Integer. Number of clusters}

\item{model}{Character. Either "poisson" or "binomial"}

\item{m}{Positive integer. Number of trials in binomial. If poisson, keep as m = 1.}
}
\value{
Matrix of posterior samples
}
\description{
Sample from posterior under RVB1
}
