% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_triangular_funcs.R
\name{make_lower_tri_matrix}
\alias{make_lower_tri_matrix}
\title{Matrix from lower triangular entries vector}
\usage{
make_lower_tri_matrix(n, omega)
}
\arguments{
\item{n}{Integer. Size of the output matrix}

\item{omega}{Numeric vector of size n(n+1)/2. Entries to make matrix.}
}
\value{
An n x n lower triangular matrix with \code{omega} entries column-wise
}
\description{
Construct a lower triangular matrix from a vector of entries
}
